# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  message: string;
}