# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyStatus } from "./propertyStatus";

export interface Property {
  id: number;
  title: string;
  slug: string;
  location: string;
  price: string;
  sqft: string;
  status: PropertyStatus;
  image: string;
  isFeatured: boolean;
  /** @nullable */
  description?: string | null;
  createdAt: Date;
  updatedAt: Date;
}