# lib/api-zod/src/generated/types/propertyInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyInputStatus } from "./propertyInputStatus";

export interface PropertyInput {
  title: string;
  slug: string;
  location: string;
  price: string;
  sqft: string;
  status: PropertyInputStatus;
  image: string;
  isFeatured?: boolean;
  /** @nullable */
  description?: string | null;
}