# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyUpdateStatus } from "./propertyUpdateStatus";

export interface PropertyUpdate {
  title?: string;
  slug?: string;
  location?: string;
  price?: string;
  sqft?: string;
  status?: PropertyUpdateStatus;
  image?: string;
  isFeatured?: boolean;
  /** @nullable */
  description?: string | null;
}