# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  id: number;
  officeAddress: string;
  phone: string;
  email: string;
  whatsappNumber: string;
  updatedAt: Date;
}