# lib/api-zod/src/generated/types/siteSettingsUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettingsUpdate {
  officeAddress?: string;
  phone?: string;
  email?: string;
  whatsappNumber?: string;
}